[backends] Make backends inherit from common base class
